backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
google-auth
google-auth-oauthlib
google-api-python-client
supabase==2.3.4
python-jose
google-generativeai